Redesigned kernel connector
